fix: Update TCP connection timeout to 30 seconds (#94)
## Which problem is this PR solving?

Old TCP stream were not being closed correctly. This was because the
timeout was set to 24 hours which lead to the number of running go
routine as we start multiple per open connection.

This PR updates the timeout so connections are flushed more frequently.

- Closes #85 

## Short description of the changes
- Remove the close timeout config option
- Sets connection timeout to 30 seconds
- Updates the assembler to use `FlushCloseOlderThan`

## How to verify that this has the expected result
Connections are now flushed and closed correctly which helps keep open
go routines in check.