fix: remove fields for http body and headers (#113)
## Which problem is this PR solving?

Disable recording of http body and header. We can consider config
options to allow users to opt-in if they choose to enable it... and/or
otherwise just be more granular about the specific data that seems
relevant.

updates (/closes?) #112 

## Short description of the changes

remove fields from telemetry:
- `http.request.body`
- `http.request.headers`
- `http.response.body`
- `http.response.headers`

## How to verify that this has the expected result

We still get other fields we've specified like `http.request.body.size`
based on the `"Content-Length"` header. But we no longer just grab the
entire header/body.