fix: Close HTTP request & response body readers (#195)
## Which problem is this PR solving?
We are not closing parsed HTTP request or response correctly. We don't
use them but not closing the reader correctly can lead to memory leaks
as the underlying byte buffer is not released and garbage collected.

## Short description of the changes
- Close both HTTP request and response body readers

## How to verify that this has the expected result
No functional changes as we're not using the HTTP request or response
bodies. This should help reduce the amount of memory creep from byte[]'s
that are not cleaned up correctly.