inject release version into image build
Adds RELEASE_VERISON as an optional build argument to the Dockerfile.

RELEASE_VERSION is still computed by the Makefile.

For the docker-build make target, pass in the version with --build-arg.

For the docker/build-and-publish GitHub Action in our Release workflow,
compute the RELEASE_VERSION with our make version target, set it in the
environment for the workflow per GitHub docs[1], and pass in the arg
with the build-args: key for the action.

[1] https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#environment-files