maint: Consolidate dockerfiles and update deps (#222)
## Which problem is this PR solving?
We maintain two Dockerfiles, one for test and one for building and
tagging. This has a maintenance burden of updating both dockerfiles
whenever changes are needed.

This PR consolidates both Dockerfiles and uses build stages for test and
build so we can target what stage we want.

## Short description of the changes
- Consolidates Dockerfile Dockerfile.test by adding test build stage
- Remove unused clang and llvm dependencies from base stage
- Update Makefile docker-build and docker-test targets to set the
desired docker build stage

## How to verify that this has the expected result
We now only have one dockerfile we have to maintain and both both build
and test makefile targets use the same one.