add a waitgroup to coordinate shutdown
- followup to #230
- closes #228
- closes #237

Use a sync.WaitGroup to ensure that both an assembler and an event
handler have stopped generating data before clean-up and process exit.

Behind the curtain, the assembler is using the libhoney client managed
within the event handler to ship assembler statistics. Closing that
libhoney client has moved from the case statement in Start() to a
separate Close() function on the event handler that is called after the
wait group count indicates that all our internal routines have ended
their loops.

How much of this interface is kept in an OTLP-flavored event handler
(and stats producer?) remains TBD.