fix: Give handler test time to process event (#242)
## Which problem is this PR solving?
There is a timing bug in the handler test where it occasionally doesn't
complete processing the event before it's closed. This makes the test
fail as libhoney didn't receive an event.

## Short description of the changes
- Add a time.sleep after adding an event to the event channel to give
the handler time to process it before starting shutdown
- Remove redundant `defer done()` as we intentionally call `done()` as
part of our shutdown processing

## How to verify that this has the expected result
The test no longer intermittently fails.