fix: skip key validation if endpoint is different (#265)
## Which problem is this PR solving?

- Closes #264 

## Short description of the changes

- Check endpoint and if it differs from our default, skip the api key
validation
- Move the Config Validate below the logging of config values; that way,
if there is something wrong, the config values are printed out first for
easier troubleshooting if it fails.

I'd considered some refactoring to pull out defaults and use a variable
instead of hardcoding here, but wasn't sure it was worth the extra
cleanup for this rare use case.

## How to verify that this has the expected result

Set an incorrect API Key. With no change to endpoint, or endpoint set to
`"https://api.honeycomb.io"`, agent pod should fail to start because of
invalid key - and should log the config options followed by the fatal
error. With a different endpoint, agent should start with no failure.