maint: Refactor packet processing to allow different event types (#271)
## Which problem is this PR solving?
Currently event processing can only deal with HTTP events. This is
inflexible and we would like to explore other event types

This PR updates the event processing to be based on a generic interfaces
for parsers and event types. This work leads the way to implement other
event type capture in the future.

## Short description of the changes
- Add `Event` interface and update `HttpEvent` to use it
- Add `Parser` interface and implement `HttpParser` using the code from
`tcpReader`
- Remove TcpReader
- Update main, handlers and tests to use new interfaces
- Update libhoney handler to separate http specific event fields and
only apply when processing a http event

## How to verify that this has the expected result
No external changes and unit tests continue to pass.

---------

Co-authored-by: Robb Kidd <robbkidd@honeycomb.io>
Co-authored-by: Jamie Danielson <jamieedanielson@gmail.com>