feat: Enable HTTP URL recording by default (#281)
## Which problem is this PR solving?
Now we strip query string parameters from the Url in #269, we can more
safely default to recording HTTP URL. The option is still be available
to opt-out if desired.

- Closes #280

## Short description of the changes
- Change default value of IncludeRequestURL config option to `true`
- Update unit tests 

## How to verify that this has the expected result
The agent's default behaviour to to record HTTP request URLs. The can be
disabled via setting the config option to false.