fix: add headers to honeycomb event (#289)
## Which problem is this PR solving?

- Closes #288 

## Short description of the changes

- add fields for `http.request.headers` and `http.response.headers` and
accompanying unit tests. #277 added all the logic for configuring the
`HTTP_HEADERS` to be extracted, and this PR adds the headers to the
actual event to be sent to Honeycomb.

This is the smaller implementation mentioned in the issue. We could
(/should?) have individual attributes for each key, but I figured this
could be a faster fix to start since we've already published with the
expectation of HTTP_HEADERS working.

## How to verify that this has the expected result

Run the agent with something like the example greeting service and see
the attributes in Honeycomb.

The default captures just the User-Agent:

- `http.request.headers`
  - example: `"{\"User-Agent\":[\"Go-http-client/1.1\"]}"`
- `http.response.headers`
  - example: `"{}"`

If for example the list includes
`"User-Agent,Accept-Encoding,Content-Length,Content-Type"`, the result
looks like this:

- `http.request.headers`
- example:
`"{\"Accept-Encoding\":[\"gzip\"],\"User-Agent\":[\"Go-http-client/1.1\"]}"`
- `http.response.headers`
- example:
`"{\"Content-Length\":[\"12\"],\"Content-Type\":[\"text/plain;
charset=utf-8\"]}"`