test the HTTP attrs added to spans
This change got more invasive than I initially intended. I extracted
most of the attribute setting to a function that returns a slice of
attributes. This avoids needing to create a span and retrieve it from
some test exporter to assert the which attributes are produced by which
logic branches.

The test is not yet comprehensive of those branches, though.