use collector IP in smoke tests
For some reason, the agent pod won't resolve DNS name for the collector
k8s service. (Maybe because it's a daemonset pod on the host network?
Different DNS available to it?)

So we need to look up the collector service's IP and use that in the
agent's Helm install values. That's done in a Make function
get_collector_ip that gets called inline to set an env var for envsubst.

BONUS! Conditionally build the docker image if it doesn't exist when
smoke testing. This necessitated breaking up the steps of smoke into
separate targets that all get set as prereqs of smoke.

Co-authored-by: Jamie Danielson <jamiedanielson@honeycomb.io>