add bats test for network agent.
WIP to limlit the test to only the echoserver 404 span.
All scopes have the same hny-network-agent name so it is difficult
to target only a specific resourceSpan for this request.
For now there is a traces-temp-test.json file being used that contains
only the resource span for the echo server (copied from the log output).
Tests check for base requirements like span ID and http attributes,
but also includes custom stuff like resource attributes and headers.